reove cnf mapper
I moved this into the claims mapper
