+++ containers/keycloak-protocol-mapper/custom-mapper/src/main/java/com/virtru/keycloak/DPoPPayload.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.keycloak.jose.jwk.JWK;
 
 import java.io.IOException;
