+++ containers/keycloak-protocol-mapper/custom-mapper/src/main/resources/META-INF/services/org.keycloak.protocol.ProtocolMapper
@@ -1,2 +1,1 @@
-com.virtru.keycloak.DPoPConfirmationMapper
 com.virtru.keycloak.TdfClaimsMapper
